Refactored handling of Positions property in Excel Implementation
Updated the way the Positions property of a BomLine object is handled. Instead of a single string, it is now stored as a List of strings; this was achieved by using the Split function to separate the string data into list items based on the comma delimiter. This change was propagated to the SylvanImplementation, NPOI and Utilities also to respect the new data structure. This assures consistency in how Positions are handled across different implementations of the Excel reader/writer. Also, ExcelColumns.cs was renamed for simplification.

In SylvanExcel implementation, instead of directly returning records, parsing to BomLine object was added which supports the changed Positions property. Its WriteBom function is implemented instead of being left as NotImplementedException.